apps/api/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
python-dotenv==1.2.2
alembic==1.18.4
psycopg2-binary==2.9.11
pydantic[email]==2.12.5
PyJWT[crypto]==2.10.1
httpx==0.28.1
resend==2.23.0
mangum==0.21.0
slowapi==0.1.9
boto3==1.42.68
fpdf2==2.8.3
